+++ NEWS.md
@@ -6,6 +6,7 @@
 * Enhanced `tm_t_summary_by` module to have geometric mean in the encoding panel.
 * Updated and added a footnote to `tm_t_summary` module.
 * Enhanced `tm_t_summary` module to have geometric mean in the encoding panel.
+* Updated `tm_t_binary_outcome` to enable an option to apply a continuity correction in the Newcombe method
 
 ### Bug fixes
 * Updated the validation to account for the error when multiple variables are not selected in `tm_g_pp_patient_timeline` module.