+++ NEWS.md
@@ -5,6 +5,7 @@
 * Updated the `tm_g_pp_adverse_events` plot to render the color legend. 
 * Enhanced `tm_t_summary_by` module to have geometric mean in the encoding panel.
 * Updated and added a footnote to `tm_t_summary` module.
+* Updated `tm_t_summary` to display the checkboxes for numeric variables statistics when numeric variables are present in the choices only.
 
 ### Bug fixes
 * Updated the validation to account for the error when multiple variables are not selected in `tm_g_pp_patient_timeline` module.