+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.modules.clinical 0.8.12.9007
+# teal.modules.clinical 0.8.12.9008
 
 ### Enhancements
 * Updated the position of the labels in the `tm_g_pp_adverse_events` plot.
@@ -7,6 +7,7 @@
 * Updated and added a footnote to `tm_t_summary` module.
 * Enhanced `tm_t_summary` module to have geometric mean in the encoding panel.
 * Updated `tm_t_binary_outcome` to enable an option to apply a continuity correction in the Newcombe method
+* Updated `tm_t_summary` to display the checkboxes for numeric variables statistics only when numeric variables are part of the selected.
 
 ### Bug fixes
 * Updated the validation to account for the error when multiple variables are not selected in `tm_g_pp_patient_timeline` module.